proj2 fix

+++ code/basic/project2/data/matrix_process.cpp
@@ -7,7 +7,7 @@
 using namespace std;
 
 int main(){
-    int matSize[] = {32, 256, 2048};
+    size_t matSize[] = {32, 256, 2048};
 
     for (int COL_NUM: matSize){
         cout << COL_NUM << ":";
@@ -17,18 +17,18 @@ int main(){
         double * matRes = new double[COL_NUM*COL_NUM]();
 
         auto start = chrono::high_resolution_clock::now();
-        for (int i = 0; i < COL_NUM; i++){
-            for (int j = 0; j < COL_NUM; j++){
+        for (size_t i = 0; i < COL_NUM; i++){
+            for (size_t j = 0; j < COL_NUM; j++){
                 matA[i*COL_NUM + j] = double(rand())/RAND_MAX;
                 matB[i*COL_NUM + j] = double(rand())/RAND_MAX;
             }
         }
 
         // 计算结果
-        for (int i = 0; i < COL_NUM; i++){
-            for(int j = 0; j < COL_NUM; j++){
+        for (size_t i = 0; i < COL_NUM; i++){
+            for(size_t j = 0; j < COL_NUM; j++){
                 double temp = 0;
-                for (int l = 0; l < COL_NUM; l++){
+                for (size_t l = 0; l < COL_NUM; l++){
                     temp += matA[i*COL_NUM + l] * matB[l*COL_NUM + j];
                 }
                 matRes[i*COL_NUM + j] = temp;
@@ -39,8 +39,8 @@ int main(){
         ofstream matBFile("mat-B-" + to_string(COL_NUM) + ".txt");
         ofstream matResFile("out" + to_string(COL_NUM) + ".txt");
 
-        for (int i = 0; i < COL_NUM; i++){
-            for (int j = 0; j < COL_NUM; j++){
+        for (size_t i = 0; i < COL_NUM; i++){
+            for (size_t j = 0; j < COL_NUM; j++){
                 matAFile << fixed << setprecision(15) << matA[i*COL_NUM + j] << " ";
                 matBFile << fixed << setprecision(15) << matB[i*COL_NUM + j] << " ";
                 matResFile << fixed << setprecision(15) << matRes[i*COL_NUM + j] << " ";
